fix(parser): fix conditional comments pushing elements into tree
The lexer would yield tokens for all the content even when inside a comment (unlike a regular non-ie browser).
This was useful to validate the markup inside the condition but would trigger false positives with the content model.

Until branching support in the DOM tree is present the content inside the comment should be ignored.

Fixes #51.