feat(htmlvalidate): retain `offset` when yielding multiple sources
If a transformer yields multiple sources the original `offset` should not be
reset to 0 for each one but rather retain the original offset similar to
lines/columns. Since the original caller has no way to know where the
transformer has yielded sources a user relying on `offset` has no way of knowing
the correct positions if it is reset.