feat(meta): extend element attribute metadata
fixes #71

BREAKING CHANGE: The metadata `attribute` has changed from `string[]` to
`MetaAttribute[]`. This primarly affects API usage.

Files read with the old format will still be supported (migrated on-the-fly) but
is deprecated and support will be removed in a future version. Typescript
declarations will not support the deprecated format and must be migrated.