chore: migrate to `dist` folder
BREAKING CHANGE: The `build` folder has been renamed to `dist`.

This affects API users only and in general should not be an issue unless
importing files via full path. In that case replace `import
'html-validate/build/...'` with `import 'html-validate/dist/...` but in general
those imports are discouraged.

Instead users should import only via `import { ... } from "html-validate"` and file an issue
if an export is missing.

This does not affect the `elements` imports which is considered a safe to import
as-is.