feat(api)!: add `Promise` based API to `HtmlValidate` class
BREAKING CHANGE: The `HtmlValidate` class now has a `Promise` based API where most
methods return a promise. The old synchronous methods are renamed.

Either adapt to the new asynchronous API:

```diff
-const result = htmlvalidate.validateFile("my-awesome-file.html");
+const result = await htmlvalidate.validateFile("my-awesome-file.html");
```

or migrate to the synchronous API:

```diff
-const result = htmlvalidate.validateFile("my-awesome-file.html");
+const result = htmlvalidate.validateFileSync("my-awesome-file.html");
```

For unittesting with Jest it is recommended to make the entire test-case async:

```diff
-it("my awesome test", () => {
+it("my awesome test", async () => {
   const htmlvalidate = new HtmlValidate();
-  const report = htmlvalidate.validateString("...");
+  const report = await htmlvalidate.validateString("...");
   expect(report).toMatchCodeFrame();
});
```