feat(api)!: new `getContextualDocumentation` to replace `getRuleDocumentation`
BREAKING CHANGE: A new `getContextualDocumentation` replaces the now deprecated
`getRuleDocumentation` method. The context parameter to `getRuleDocumentation`
is now required and must not be omitted.

For rule authors this means you can now rely on the `context` parameter being
set in the `documentation` callback.

For IDE integration and toolchain authors this means you should migrate to use
`getContextualDocumentation` as soon as possible or if you are continuing to use
`getRuleDocumentation` you are now required to pass the `config` and `context`
field from the reported message.