fix(rules): ignore parents of template tags when checking permitted descendants
Elements within a `<template>` tag are not children of the `<template>`
according to the [HTML specifications](https://html.spec.whatwg.org/#the-template-element).

When checking if a node is a valid descendant of a ancestor, `element-permitted-content`
now stops as soon as it encounters a `<template>` tag when going up the node's parent list.
This means ancestors of the `<template>` tag are now rightfully ignored
when deciding if a node is permitted as a descendant.

Fixes #305